refactored rocket node tests
